WebService stuff 09.04.20
Added data.csv with which relevant data the experiment is running.
Bash_history is added from the Ubuntu terminal to be able to see how the solution is installed as a WebService, as well as links to instructions on how to set up the WebService on Ubuntu OS. Container will be provided later.